End of Part 4

+++ controllers/blogs.js
@@ -2,21 +2,35 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
+const jwt = require('jsonwebtoken')
+
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('author', { username: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
+const getTokenFrom = request => {
+    const authorization = request.get('authorization')
+    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+        return authorization.substring(7)
+    }
+    return null
+}
+
 blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
-
-    const user = await User.findById(body.authorId)
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
 
     const blog = new Blog({
         title: body.title,
         url: body.url,
         likes: body.likes,
-        author: user._id
+        authorId: user._id
     })
 
     const savedBlog = await blog.save()
@@ -26,9 +40,20 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    console.log(request.params.id)
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const token = getTokenFrom(request)
+
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if (decodedToken.id.toString() === user._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        return response.status(204).end()
+    }
+    return response.status(401).json({ error: 'not authorized to perform this request' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -36,7 +61,7 @@ blogsRouter.put('/:id', async (request, response) => {
 
     const blog = {
         title: body.title,
-        author: body.author,
+        authorId: body.authorId,
         url: body.url,
         likes: body.likes
     }
